+++ packages/expo/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### 💡 Others
 
+- Remove `AppRegistry.setWrapperComponentProvider` in favor of `registerRootComponent`.
 - Add `@expo/config-plugins` dependency for packages that have a peer dependency on `expo`. ([#18595](https://github.com/expo/expo/pull/18595) by [@EvanBacon](https://github.com/EvanBacon))
 - Convert `DevAppContainer` to functional React component. ([#18597](https://github.com/expo/expo/pull/18597) by [@EvanBacon](https://github.com/EvanBacon))